text limit functional in each template && upload cropping functional for imagewidget. Removed template 5
