+++ src/components/student/bookTemplates/index.js
@@ -2,21 +2,18 @@ import Template1 from "./template1";
 import Template2 from "./template2";
 import Template3 from "./template3";
 import Template4 from "./template4";
-import Template5 from "./template5";
 
 // Descriptions of templates.
 // T1: all text
 // T2: one full page image
-// T3: top right image
-// T4: top left image
-// T5: full screen image with title
+// T3: top image bottom text
+// T4: top text bottom image
 
 const templates = {
     Template1,
     Template2,
     Template3,
-    Template4,
-    Template5
+    Template4
 }
 
 export default templates;
