+++ src/components/student/bookTemplates/template2.js
@@ -6,7 +6,9 @@ const Template2 = () => {
         <div className="page-outer-div temp2-outer-div">
             <div className={"image-center "}>
                 <ImageWidget 
-                    croppingRatio={0.9}
+                    croppingRatio={0.57916666666}
+                    maxHeight={480}
+                    maxWidth={278}
                 />
             </div>
         </div>
