+++ src/components/student/bookTemplates/template4.js
@@ -1,47 +1,32 @@
-import React, { useState, useRef } from 'react';
+import React, { useState, useEffect } from 'react';
 import ImageWidget from './uploadWidget';
+import rowLimiter from './rowLimiter';
 
-const Template4 = () => {
-    const [text, setText] = useState('');
-    const pageText = useRef(null);
+const Template4 = ({ page }) => {
+    const [text, setText] = useState("");
 
-    function textSetter(){
-        setText(pageText.current.innerHTML);
-    }
+    useEffect(() => {
+        if (page.id) {
+            setText(page.content || "");
+        }
+    }, [page])
 
     return (
         <div className="page-outer-div temp4-outer-div">
-            <div className='text-img-div'>
+            <div className='text-div'>
+                <textarea className="full-text-page" defaultValue={text } onChange={(event) => setText(event.target.value) }
+                    rows="10" cols="30"
+                    spellCheck="true"
+                    onKeyPress={rowLimiter(event, 10)}
+                ></textarea>
                 <div className="small-image-page">
                     <ImageWidget 
-                        croppingRatio={0.743}
-                        maxHeight={175}
-                        maxWidth={130} 
-                        minWidth={130}
+                        croppingRatio={1.54444444444}
+                        maxHeight={180}
+                        maxWidth={278}
                     />
                 </div>
-                <div className="small-text-page" ref={pageText} contentEditable="true">
-                    <br></br><br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                    <br></br>
-                </div>
-            </div>        
+            </div>
         </div>
     );
 };
