+++ src/less/pageTemplates/template2.less
@@ -6,11 +6,11 @@
         display: flex;
         align-items: center;
         justify-content: center;
-        margin: 5px 0 0 15px;
-        width: 100%;
-        height: 370px;
-        margin-right: 20px;
-        border: dashed  2px;
-        z-index: 1;    
+        margin: 5px;
+        min-width: 278px;
+        max-width: 278px;
+        min-height: 480px;
+        max-height: 480px;
+        background-color: rgb(243, 243, 243); 
     }
 }
