+++ src/less/pageTemplates/template3.less
@@ -1,30 +1,31 @@
 @import "../common.less";
 
 .temp3-outer-div{
-    flex-direction: column;
-    padding-top: 10px;
+    
+    .text-div{
+        height: 100%;
+        justify-content: space-between;
+        margin: 5px 10px;
 
-    .text-img-div{
-        width: 100%;
-        
         .small-image-page{
-            float: right;
+            min-height: 180px;
+            max-height: 180px;
+            background-color: rgb(240, 240, 240);
             display: flex;
             justify-content: center;
             align-items: center;
-            min-width: 130px;
-            max-width: 130px;
-            min-height: 175px;
-            max-height: 175px;
-            margin: 0 5px 5px 10px;
-            border: dashed rgb(245, 227, 136) 2px;
         }
-    
-        .small-text-page{
-            margin-left: 10px;
-            max-width: 92%;
-            max-height: 90%;
-        }
-}
 
+        .full-text-page{
+            margin-top: 5px;
+            min-width: 278px;
+            max-width: 278px;
+            min-height: 278px;
+            max-height: 278px;
+            font-size:x-large;
+            overflow: hidden;
+            resize:none;
+            border: none;
+        }
+    }
 }
