+++ src/less/pageTemplates/template4.less
@@ -1,30 +1,32 @@
 @import "../common.less";
 
 .temp4-outer-div{
-    max-height: 100%;
     
-    .text-img-div{
-        width: 100%;
-        
+    .text-div{
+        height: 100%;
+        justify-content: space-between;
+        margin: 5px 10px;
+
         .small-image-page{
-            float: left;
+            min-height: 180px;
+            max-height: 180px;
+            background-color: rgb(240, 240, 240);
             display: flex;
             justify-content: center;
             align-items: center;
-            min-width: 130px;
-            max-width: 130px;
-            min-height: 175px;
-            max-height: 175px;
-            margin: 0 10px 5px 5px;
-            border: dashed rgb(245, 227, 136) 2px;
+            margin-top: 5px;
         }
-    
-        .small-text-page{
-            margin-left: 10px;
-            max-width: 95%;
-            min-height: 90%;
-            max-height: 90%;
-            
+
+        .full-text-page{
+            min-width: 278px;
+            max-width: 278px;
+            min-height: 278px;
+            max-height: 278px;
+            font-size:x-large;
+            overflow: hidden;
+            resize:none;
+            border: none;
         }
     }
 }
+
